+++ protocols/request-response/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.26.0 - unreleased
+
+- Remove `request_response::Config::set_connection_keep_alive` in favor of `SwarmBuilder::idle_connection_timeout`.
+  See [PR 4679](https://github.com/libp2p/rust-libp2p/pull/4679).
+
 ## 0.25.2
 
 - Deprecate `request_response::Config::set_connection_keep_alive` in favor of `SwarmBuilder::idle_connection_timeout`.